# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Question API swagger documentation",
        "contact": {
            "email": "echo.zahar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.2"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Demo tech task API server"
        }
    ],
    "paths": {
...
        "/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Получение всех обращении с использоваение простой пагинаций, по 3 обращения (simplePaginate(3))",
                "operationId": "QuestionAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Пагинация по страницам",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно загружено."
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Сохранение новой заявки от пользователя",
                "description": "Store a newly created resource in storage.",
                "operationId": "storeQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Обращение успешно добавлено.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questions"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/questions/{id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Показать запись по id",
                "description": "Display the specified resource.",
                "operationId": "getQuestionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные успешно, загружены.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Questions"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Что то, пошло не так, записи не нaйдено."
                    }
                }
            },
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Обновление записи. Добавить ответ от администратора",
                "description": "Update the specified resource in storage.",
                "operationId": "questionUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Метод обновления, по id записи.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Запись успешно обновлена"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "не авторизован"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Что то пошло не так, запись не найдена."
                    },
                    "419": {
                        "description": "Что то пошло не так!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Мягкое удаление (из БД не удаляется) обращения",
                "description": "Remove the specified resource from storage.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Введите id записи для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Запись успешно удалена"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Что то пошло не так, запись не найдена."
                    }
                }
            }
        },
        "/questions/search/{name}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "поиск по имени пользователя составивщего заявку",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "поиск по имени пользователя составивщего заявку",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "не найдено"
                    },
                    "200": {
                        "description": "статус 'ОК'"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "QuestionStoreRequest": {
                "title": "Сохранение обращения пользователя.",
                "description": "создание новой записи, заявка от пользователя.",
                "properties": {
                    "name": {
                        "title": "имя пользователя",
                        "description": "Нужно ввести имя пользователя. Обязательное поле, минимальная длина 2, максимальное 50.",
                        "type": "string",
                        "format": "string",
                        "example": "Customer"
                    },
                    "email": {
                        "title": "email пользователя.",
                        "description": "email пользователя. Обязателеное поле, минимальная длина 5, максимальная 100. ",
                        "type": "string",
                        "format": "string",
                        "example": "customer@example.com"
                    },
                    "message": {
                        "title": "сообщение от пользователя.",
                        "description": "сообщение, обращение пользователя, обязателеное поле с максимальным значением 500.",
                        "type": "string",
                        "format": "string",
                        "example": "Мое сообщение ..........."
                    }
                },
                "type": "object"
            },
            "QuestionUpdateRequest": {
                "title": "Ответ пользователю от администратору.",
                "description": "Обновление записи, заявка от пользователя",
                "properties": {
                    "comment": {
                        "title": "Ответ пользователю от администратора.",
                        "description": "Введите ответ пользователю, минимальная длина 5,  максимальная 500 символов.",
                        "type": "string",
                        "format": "string",
                        "example": "какой то, текст."
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "Авторизированный пользователь",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Questions": {
                "title": "Обращения",
                "description": "Модель обращении пользователей, структура и описание.",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id обрщения",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Имя пользователя",
                        "description": "имя обращающегося",
                        "type": "string",
                        "format": "string",
                        "example": "customer"
                    },
                    "email": {
                        "title": "email обращающегося",
                        "description": "email пользователя (добавить описане из валидаций)",
                        "format": "email",
                        "example": "customer@example.com"
                    },
                    "status": {
                        "title": "статус записи",
                        "description": "2 вида статуса: active (активные обращения) и resolved - закрытый",
                        "format": "email",
                        "example": "active"
                    },
                    "message": {
                        "title": "Обращение от пользователя",
                        "description": "",
                        "format": "string",
                        "example": "Это пример текста обращения."
                    },
                    "dateAction": {
                        "title": "Дата ответа от админа",
                        "description": "Поле не обязательно для заполнения",
                        "format": "datetime",
                        "example": "2021-12-21 21:21:21"
                    },
                    "comment": {
                        "title": "Ответ администратора",
                        "description": "Ответ администратора на обращение пользователя",
                        "format": "string",
                        "example": "Ответ на ваш вопрос ...."
                    },
                    "user_id": {
                        "title": "ID польлзователя (админ)",
                        "description": "id пользователя ответившего на обращение.",
                        "type": "integer",
                        "format": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "title": "создание обращения",
                        "description": "Дата создания обращения пользователя",
                        "format": "datetime",
                        "example": "2021-12-21 21:21:21"
                    },
                    "updated_at": {
                        "title": "обновление записи обращения",
                        "description": "дата последнего обновления записи.",
                        "format": "datetime",
                        "example": "2021-12-21 21:21:21"
                    },
                    "deleted_at": {
                        "title": "удаление обращения",
                        "description": "Мягкое удаление, записывается дата удаления обращения",
                        "format": "datetime",
                        "example": "2021-12-21 21:21:21"
                    },
                    "user": {
                        "title": "администратор",
                        "description": "пользователь - администратор, который пищет ответ на обращение."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Question"
                }
            }
        },
        "securitySchemes": {
            "TOKEN": {
                "type": "apiKey",
                "description": "Этот контроллер создан исключительно для документаций и ни где не будет использоваться в проекте.",
                "name": "TOKEN",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}